% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akalbers_proj.R
\docType{data}
\name{akalbers}
\alias{akalbers}
\title{Alaska Albers Projection}
\format{
An object of class \code{character} of length 1.
}
\usage{
akalbers
}
\description{
"+proj=aea +lat_1=55 +lat_2=65 +lat_0=50 +lon_0=-154 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"  # make object of the code defining the desired projection (in this case Alaska Equal Area Albers [ EPSG: 3338])
}
\keyword{datasets}
