% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocruncher.R
\name{autocruncher}
\alias{autocruncher}
\title{autocruncher -- find all the bad knots
 based on soap.r in mgcv
 Copyright Simon Wood 2006-2012}
\usage{
autocruncher(bnd, knots, nmax = 200, k = 10, xname = "x", yname = "y")
}
\arguments{
\item{bnd}{boundary}

\item{knots}{knotd}

\item{nmax}{nmax, default=200}

\item{k}{boundary basis functions, default =10}

\item{xname}{x}

\item{yname}{y}
}
\value{

}
\description{
bugs added by David L Miller, 2019
}
\details{
NB in mgcv nmax is 100 here, where as the default is 200
}
