% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad_knots_lp.R
\name{bad_knots_lp}
\alias{bad_knots_lp}
\title{find all the bad knots with DSM}
\usage{
bad_knots_lp(knots, dsmk, nmax = 200)
}
\arguments{
\item{knots}{dataframe of knots used in soap_check}

\item{k}{k to be used in DSM}
}
\value{
dataframe of knots without the ones identified as "on or outside bounds" by dsm
}
\description{
find all the bad knots with DSM
}
