% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHAPODAS.R
\name{convertHAPODAS}
\alias{convertHAPODAS}
\title{converts raw HAPODAS dataframe to work with SplitTransects functions}
\usage{
convertHAPODAS(data)
}
\arguments{
\item{data}{a raw HAPODAS dataset, column names unaltered.}
}
\value{
Returns a dataframe with specified columns:

datetime: POSIXct class, date and time of datapoint

lats: numeric class, decimal latitude of datapoint

longs: numeric class, decimal longitude of datapoint

label: factor class, unique line label for transect

onoff: factor class, on or off effort designation for line

section: factor class, section of line

travel: numeric class, distance traveled since last data point IN METERS, converted from Nautical miles by multiplying by 1852 (m/nm)

tdiff: numeric class, time difference in seconds to next point

orriginal: Logical class, designates row as an orriginal datapoint
}
\description{
converts raw HAPODAS dataframe to work with SplitTransects functions
}
\author{
Abigail Schiffmiller
}
