% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_by_group.R
\name{cumsum_by_group}
\alias{cumsum_by_group}
\title{Cumulative sum by grouping variables}
\usage{
cumsum_by_group(x, grpVarLst)
}
\arguments{
\item{x}{vector (numeric or logical). If x is a column in a dataset, dataset must be identified (e.g: dataset$x)}

\item{grpVarLst}{a `list` of grouping variables each of same length as x. If in dataset, dataset must be identified.}
}
\value{
Returns a vector of the same length as x. Each unique combination of grouping variables has it's own cumulative sum.
}
\description{
Runs cumsum on x seperately for each unique combination of the grouping variables
}
\author{
Abigail Schiffmiller
}
