% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distdata.R
\name{Distdata}
\alias{Distdata}
\title{converts raw HAPODAS dataframe to DistData data for DSM and detection function}
\usage{
Distdata(data, covariates)
}
\arguments{
\item{data}{a raw HAPODAS dataset, column names unaltered.}

\item{covariates}{a list of covariate column names}
}
\value{
Returns a Dataframe or SpatialPointsDataFrame (if makeSP=T) with one row per sighting with specified columns:

datetime: POSIXct class, timestamp of datapoint, links to obsdata

lats: numeric class, decimal latitude of datapoint, links to obsdata

longs: numeric class, decimal longitude of datapoint, links to obsdata

label: factor class, unique line label for transect, links to obsdata

distance *** : REQUIRED for package `Distance`: the perpendicular distance from transect line to the detection (from HAPODAS$perpDist_km)

object *** : REQUIRED for package `Distance`: unique object identifier (from HAPODAS$SIGHTNUMBER), links to obsdata

size *** : REQUIRED for package `Distance`: group size

species: SPECIESCODE

detected: *required for mrds* 1 if detected by the observer and 0 if missed **(always 1 for single observer)**

observer : *required for mrds* observer number (1 or 2) **(always 1 for single observer)**

other factor covariates used for detection function

x: (only after running `makeSP`) numeric class, the UTM (eastings) reprojection of longitude, default CRS is Alaska Equal Area Albers

y: (only after running `makeSP`) numeric class, the UTM (northings) reprojection of latitude, default CRS is Alaska Equal Area Albers

'***' indicates a required item for this dataframe for `dsm` or `Distance`
}
\description{
converts raw HAPODAS dataframe to distance data: Holds detection data including ID, species, distance, covariates for detection function
}
\author{
Abigail Schiffmiller
}
