% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbwDSMdata2km.R
\docType{data}
\name{hapodas_Segdata}
\alias{hapodas_Segdata}
\title{hapodas Segdata
made with `Segdata` function. All hapodas segments}
\format{
An object of class \code{data.frame} with 1670 rows and 9 columns.
}
\usage{
hapodas_Segdata
}
\description{
hapodas Segdata
made with `Segdata` function. All hapodas segments
}
\keyword{datasets}
