% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbwDSMdata2km.R
\docType{data}
\name{hbw_Dist}
\alias{hbw_Dist}
\title{Data for HBW DSM vignette, 2km segments
data constucted in file: "~/UAF/MS/data and analysis/HAPODAS/dsmdata_species.Rmd"}
\format{
An object of class \code{data.frame} with 395 rows and 18 columns.
}
\usage{
hbw_Dist
}
\description{
usethis::use_data(hbw_Dist, hbw_Obs, hapodas_Segdata, hbw_hr_cds.05, hbw_hr_mcds_CUE_SIGHTOBS, overwrite=TRUE)
}
\details{
HBW Distdata
made with `Distdata` function. Only Humpback whale sightings.
}
\keyword{datasets}
