% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbwDSMdata2km.R
\docType{data}
\name{hbw_hr_mcds_CUE_SIGHTOBS}
\alias{hbw_hr_mcds_CUE_SIGHTOBS}
\title{HBW ddf.obj: an MCDS detection function for humpback whales with:
 hazard rate key function
 sighting cue (`CUE`) covariate
 sighting observer (`SIGHTOBS`) covariate
 the most distant 5% of sighings truncated}
\format{
An object of class \code{dsmodel} of length 3.
}
\usage{
hbw_hr_mcds_CUE_SIGHTOBS
}
\description{
hist(hbw_Dist$distance, breaks= 15)
hbw_hr_mcds_CUE_SIGHTOBS<-Distance::ds(hbw_Dist, truncation="5%", key='hr',
                                       adjustment=NULL, formula= ~CUE+SIGHTOBS)
summary(hbw_hr_mcds_CUE_SIGHTOBS)
plot(hbw_hr_mcds_CUE_SIGHTOBS, main= "Humpback whale: MCDS Hazard Rate+ CUE + SIGHTOBServer")
}
\keyword{datasets}
