% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbwDSMdata2km.R
\docType{data}
\name{hbw_Obs}
\alias{hbw_Obs}
\title{HBW Obsdata
made with `Obsdata` function. Only Humpback whale sightings.}
\format{
An object of class \code{data.frame} with 395 rows and 9 columns.
}
\usage{
hbw_Obs
}
\description{
HBW Obsdata
made with `Obsdata` function. Only Humpback whale sightings.
}
\keyword{datasets}
