% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_by_secs.R
\name{interp_by_secs}
\alias{interp_by_secs}
\title{interpolate location for every second along transect}
\usage{
interp_by_secs(dataset)
}
\arguments{
\item{dataset}{a dataframe with specified columns:

datetime: POSIXct class, date and time of datapoint

lats: numeric class, decimal latitude of datapoint

longs: numeric class, decimal longitude of datapoint

label: character or factor class, unique line label for transect

onoff: character or factor class, on or off effort designation for line

section: character or factor class, section of line

travel: numeric class, distance traveled since last data point IN METERS

tdiff: tdiff class, time difference in seconds from previous point

orriginal: Logical class, designates row as an orriginal datapoint}
}
\value{
Returns a dataframe with columns:

datetime: POSIXct class, date and time of datapoint. One row for every second

lats: numeric class, decimal latitude of datapoint. Interpolated for every second between orriginal datapoints.

longs: numeric class, decimal longitude of datapoint. Interpolated for every second between orriginal datapoints.

label: character or factor class, unique line label for transect

onoff: character or factor class, on or off effort designation for line

section: character or factor class, section of line

travel: numeric class, distance traveled since LAST data point IN METERS

orriginal: Logical class, designates row as an orriginal datapoint
}
\description{
interpolate location for every second along transect
and keep the transect metadata/identifying
info for new rows, for line transect dataset.
}
\examples{
require(tictoc)
tic()
micro_interp  <- interp_by_secs(micro)
toc()
head(micro_interp)

}
\author{
Abigail Schiffmiller
}
