% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSP.R
\name{makeSP}
\alias{makeSP}
\title{turns a dataframe into SP class}
\usage{
makeSP(dataset, coordx, coordy, proj = akalbers)
}
\arguments{
\item{dataset}{a dataframe to be converted to an SP class}

\item{coordx}{the variable of the dataframe to be used as the `x` coordinate}

\item{coordy}{the variable of the dataframe to be used as the `y` coordinate}

\item{proj}{the projection to use for the SPDF. Default is akalbers (Alaska Equal Area Albers)

#' @return Returns a SpatialPointsDataFrame version of the input dataset with coordinates defined as `x` and `y`}
}
\description{
turns a dataframe into SP class. To be used with `HAPODASdsmR` functions: `dsm_segdata`, `Distdata`, and `Obsdata` for DSM analysis
}
