% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_shortsegs.R
\name{merge_shortsegs}
\alias{merge_shortsegs}
\title{combine short end segmnets to previous}
\usage{
merge_shortsegs(data, cutoff = 1000)
}
\arguments{
\item{data}{a dataframe with specified columns:

datetime: POSIXct class, date and time of datapoint. One row for every second

lats: numeric class, decimal latitude of datapoint. Interpolated for every second between orriginal datapoints.

longs: numeric class, decimal longitude of datapoint. Interpolated for every second between orriginal datapoints.

label: character or factor class, unique line label for transect

onoff: character or factor class, on or off effort designation for line

section: character or factor class, section of line

travel: numeric class, distance traveled since LAST data point IN METERS

orriginal: Logical class, designates row as an orriginal datapoint "True" or interpolated point "FALSE"

cum_dist: numeric class, from `cumsum_by_group`, cumulative distance in meteres for each unique label/onoff/section combination

mod: numeric class, the modulus of cum_dist by seg.length

cutpoint: Character class: identifies "Start" and "End" points of each segment. Points inside a segment are designated "0"}

\item{cutoff}{segments shorther than this length will be merged with the provious segment id it has the same line and section number. Defoult value is 1000m}
}
\value{
a dataframe with specified columns:

datetime: POSIXct class, date and time of datapoint. One row for every second

lats: numeric class, decimal latitude of datapoint. Interpolated for every second between orriginal datapoints.

longs: numeric class, decimal longitude of datapoint. Interpolated for every second between orriginal datapoints.

label: character or factor class, unique line label for transect

onoff: character or factor class, on or off effort designation for line

section: character or factor class, section of line

travel: numeric class, distance traveled since LAST data point IN METERS

orriginal: Logical class, designates row as an orriginal datapoint "True" or interpolated point "FALSE"

cum_dist: numeric class, from `cumsum_by_group`, cumulative distance in meteres for each unique label/onoff/section combination

mod: numeric class, the modulus of cum_dist by seg.length

cutpoint: Character class: identifies "Start" and "End" points of each segment. Points inside a segment are designated "0"

segnum: interger class, Unique ID number for each segment

sec_label: character class, a combination of Label and section for a single identifying value
}
\description{
Reassigns 'segnum' to match previous segment if max segment length is less than specified
 AND previous shares same label and section value
}
\author{
Abigail Schiffmiller
}
