% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_data.R
\docType{data}
\name{micro}
\alias{micro}
\title{10 obs. of 9 variables}
\format{

}
\usage{
micro
}
\description{
A dataframe as produced by `convertHAPODAS` from raw HAPODAS data rows 547:556,
with "label", "onoff", and "section" edited to represent conditions of a larger
dataset better.
}
\keyword{datasets}
