% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedistdata.R
\docType{data}
\name{micro_distdata}
\alias{micro_distdata}
\title{making micro_distdata and short_distdata}
\format{

}
\usage{
micro_distdata
}
\description{
micro: 1 obs. of 13 variables
 short: 9 obs. of 13 variables
}
\details{
A dataframe of the necessary columns to do detection functions ising package `Distance`
}
\keyword{datasets}
