% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obsdata.R
\name{Obsdata}
\alias{Obsdata}
\title{makes an 'observation' dataframe}
\usage{
Obsdata(Dist_data, Seg_data)
}
\arguments{
\item{Dist_data}{Distdata dataframe from `Distdata`}

\item{Seg_data}{Segdata dataframe from `dsm_segdata`}
}
\value{
Returns a Dataframe or SpatialPointsDataFrame (if makeSP=T) with one row per sighting with specified columns:

datetime: POSIXct class, timestamp of datapoint (FROM Dist_data)

longs: numeric class, decimal longitude of datapoint (FROM Dist_data)

lats: numeric class, decimal latitude of datapoint (FROM Dist_data)

label: factor class, unique line label for transect (FROM Dist_data)

distance: the perpendicular distance from transect line to the detection (FROM Dist_data)

object: unique object identifier (FROM Dist_data)

size: group size (FROM Dist_data)

Effort: numeric class, length in meters of the segment (FROM Seg_data)

Sample.Label: character or factor class, unique segment label (FROM Seg_data)

x: (only after running `makeSP`) numeric class, the UTM (eastings) reprojection of longitude, default CRS is Alaska Equal Area Albers

y: (only after running `makeSP`) numeric class, the UTM (northings) reprojection of latitude, default CRS is Alaska Equal Area Albers
}
\description{
makes an 'observation' dataframe which links segment data with distance data
}
\author{
Abigail Schiffmiller
}
