% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segdata.R
\name{Segdata}
\alias{Segdata}
\title{Converts from 'segmentized' to 'segdata' dataframes
takes segmentized dataset with all start and end points and original datapoints and gives "segdata" data set to be used by 'dsm'}
\usage{
Segdata(data)
}
\arguments{
\item{data}{a dataframe as produced by SplitTransects::segmentize}
}
\value{
a Dataframe or SpatialPointsDataFrame (if makeSP=T) with one row per segment with following columns:

datetime: POSIXct class timestamp of datapoint

longitude: numeric class, the 'center' longitude of the segment (unless 'start','endpoint' or another point is specified) (from longs)

latitude: numeric class, the 'center' latitude of the segment (unless 'start','endpoint' or another point is specified) (from lats)

Effort: numeric class, length in meters of the segment (from mod)

Transect.Label: character or factor class, unique line label for transect (from label)

Sample.Label: character or factor class, unique segment label (combination of label and segnum)

x: (only after running `makeSP`) numeric class, the UTM (eastings) reprojection of longitude, default CRS is Alaska Equal Area Albers

y: (only after running `makeSP`) numeric class, the UTM (northings) reprojection of latitude, default CRS is Alaska Equal Area Albers


 @author Abigail Schiffmiller
}
\description{
Converts from 'segmentized' to 'segdata' dataframes
takes segmentized dataset with all start and end points and original datapoints and gives "segdata" data set to be used by 'dsm'
}
