% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_data.R
\docType{data}
\name{short}
\alias{short}
\title{150 obs. of 9 variables}
\format{

}
\usage{
short
}
\description{
A dataframe as produced by 'convertHAPODAS' from raw HAPODAS data rows 600:749.
}
\keyword{datasets}
