% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedistdata.R
\docType{data}
\name{short_distdata}
\alias{short_distdata}
\title{#'
short_hapodas <- hapodas[600:749,]}
\format{
An object of class \code{data.frame} with 9 rows and 13 columns.
}
\usage{
short_distdata
}
\description{
short_distdata <- Distdata(data=short_hapodas, covariates = list(CUE=short_hapodas$CUE,
                                                                 SIGHTOBS=short_hapodas$SIGHTOBS,
                                                                 Visibility=short_hapodas$Visibility,
                                                                 BEAUFORT=short_hapodas$BEAUFORT,
                                                                 FOGRAIN=short_hapodas$FOGRAIN))
}
\details{
usethis::use_data(short_distdata)
}
\keyword{datasets}
