% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_data.R
\docType{data}
\name{short_hapodas}
\alias{short_hapodas}
\title{150 obs. of all HAPODAS variables (AKA *RAW HAPODAS*): needed for 'Distdata()', can also be used for 'convertHAPODAS'}
\format{
An object of class \code{data.frame} with 150 rows and 86 columns.
}
\usage{
short_hapodas
}
\description{
short_hapodas <- hapodas[600:749,]
usethis:: use_data(short_hapodas)
}
\keyword{datasets}
