% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortSegments_data.R
\docType{data}
\name{shortSegments}
\alias{shortSegments}
\title{200 obs. of 12 variables}
\format{

}
\usage{
shortSegments
}
\description{
A dataframe as produced by SplitTransects from "short.rda" dataset
through `segmentize` function (see SplitTransects-vignette),
with "label", "onoff", and "section" edited to represent conditions of a larger
dataset better.
}
\keyword{datasets}
