% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap_check.R
\name{soap_check}
\alias{soap_check}
\title{Check whether a soap film smoother boundary and knots make sense
see Readme.md for details on how to use this
library(mgcv)
library(rgeos)
library(sp)
source("autocrunch.R")}
\usage{
soap_check(
  bnd,
  knots = NULL,
  data = NULL,
  plot = TRUE,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{bnd}{boundry}

\item{knots}{knots}

\item{data}{data}

\item{plot}{plot}

\item{tol}{tolerance}
}
\value{

}
\description{
Check whether a soap film smoother boundary and knots make sense
see Readme.md for details on how to use this
library(mgcv)
library(rgeos)
library(sp)
source("autocrunch.R")
}
