% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_names_data.R
\docType{data}
\name{species_names}
\alias{species_names}
\title{species names}
\format{
An object of class \code{data.frame} with 13 rows and 3 columns.
}
\usage{
species_names
}
\description{
table of `SPECIESCODE` from HAPODAS with corosponding species common names and scientific names
species_names <- read.csv("~/UAF/MS/data and analysis/HAPODAS/HAPODAS DATA/hapodas_species.csv")
#'
SPECIESCODE               SCISPNAME                  SPNAME
            44      Phocoenoides dalli          Dalls porpoise ***
            40       Phocoena phocoena        Harbour porpoise ***
            37            Orcinus orca                    Orca
           477                    <NA>           UnID porpoise
            76  Meagptera novaeangliae          Humpback whale ***
            71 Balenoptera acutorostra             Minke whale
            79                    <NA>              UnID whale
            EL          Enhydra lutris               Sea otter
            US                    <NA>               UnID seal
            EJ      Eumetopias jubatus        Stellar sea lion
            CU     Callorhinus ursinus       Northern fur seal
            MA Mirounga angustirostris  Northern elephant seal
            PV          Phoca vitulina             Harbor seal
}
\keyword{datasets}
